+++ popkorn/src/main/kotlin/cc/popkorn/core/exceptions/DefaultImplementationNotFoundException.kt
@@ -1,0 +1,3 @@
+package cc.popkorn.core.exceptions
+
+class DefaultImplementationNotFoundException(element: String, options: List<String>) : RuntimeException("Could not find default implementation of $element: ${options.joinToString()}")